Move runnable_with_examples into separate file
