+++ libcaf_test/caf/test/fwd.hpp
@@ -21,9 +21,11 @@ class context;
 class factory;
 class given;
 class nesting_error;
+class outline;
 class registry;
 class reporter;
 class runnable;
+class runnable_with_examples;
 class runner;
 class scenario;
 class scope;
